Fix definition of "blockchain" to match behavior
Technical documentation must match the empirically observed behavior of
the implementation.

This fix is also needed for honesty in advertising.  Investors and users
should not be misled to believe that the Terra ledger is "unchangeable",
that the stake-weighted controlling majority of its nodes is
"independent", or that it is not subject to the diktat of a central
authority.